pd-frames/requirements.txt
fastapi
fastapi-sqlalchemy
PyMySQL
SQLAlchemy
uvicorn[standard]
python-decouple
behave
yml
opencv-python
cython
wget
decord